Create GPS_data.ino
I used internal GPS feature to get GPS coordinates of the vehicle.
Here is the Arduino code for using GPS
This code for Get the string which is sent from sim808 and decode it to parts as we can separate Latitude, Longitude,Altitude, date, time ,no. of satellites and so on.